refactor: split into two files
