+++ packages/portal/src/components/DS4CH/DS4CHLandingHero.vue
@@ -14,7 +14,7 @@
         <!-- eslint-disable vue/no-v-html -->
         <div
           class="hero-content-text-block"
-          v-html="parseMarkdownHtml(`# ${headline}\n${text}`)"
+          v-html="parseMarkdown(`# ${headline}\n${text}`)"
         />
         <!-- eslint-enable vue/no-v-html -->
         <SmartLink
@@ -42,7 +42,7 @@
 
 <script>
   import SmartLink from '@/components/generic/SmartLink';
-  import { parseMarkdownHtml } from '@/utils/markdown.js';
+  import parseMarkdown from '@/utils/markdown/parse.js';
 
   const CSS_VARS_PRESETS = {
     small: { w: 576, h: 896, fit: 'fill', q: 100, f: 'right' },
@@ -107,7 +107,7 @@
     },
 
     methods: {
-      parseMarkdownHtml
+      parseMarkdown
     }
   };
 </script>
