+++ packages/portal/src/components/content/ContentPrimaryCallToAction.vue
@@ -9,7 +9,7 @@
     <!-- eslint-disable vue/no-v-html -->
     <div
       class="primary-cta-rich-text text-left"
-      v-html="parseMarkdownHtml(text)"
+      v-html="parseMarkdown(text)"
     />
     <!-- eslint-enable vue/no-v-html -->
     <SmartLink
@@ -26,7 +26,7 @@
 
 <script>
   import SmartLink from '../generic/SmartLink';
-  import { parseMarkdownHtml } from '@/utils/markdown.js';
+  import parseMarkdown from '@/utils/markdown/parse.js';
 
   export default {
     name: 'ContentPrimaryCallToAction',
@@ -55,7 +55,7 @@
     },
 
     methods: {
-      parseMarkdownHtml
+      parseMarkdown
     }
   };
 </script>
