+++ packages/portal/src/components/landing/LandingHero.vue
@@ -7,7 +7,7 @@
         <header class="hero-content">
           <!-- eslint-disable vue/no-v-html -->
           <div
-            v-html="parseMarkdownHtml(`# ${headline}\n${text}`)"
+            v-html="parseMarkdown(`# ${headline}\n${text}`)"
           />
           <!-- eslint-enable vue/no-v-html -->
           <SmartLink
@@ -40,7 +40,7 @@
 <script>
   import ImageWithAttribution from '@/components/image/ImageWithAttribution';
   import SmartLink from '@/components/generic/SmartLink';
-  import { parseMarkdownHtml } from '@/utils/markdown.js';
+  import parseMarkdown from '@/utils/markdown/parse.js';
 
   export default {
     name: 'LandingHero',
@@ -110,7 +110,7 @@
     },
 
     methods: {
-      parseMarkdownHtml
+      parseMarkdown
     }
   };
 </script>
