+++ packages/portal/src/utils/markdown/parse.js
@@ -1,0 +1,5 @@
+import { marked } from 'marked';
+
+export default (text) => {
+  return text ? marked.parse(text) : text;
+};
